Migrate tags page to contentful
